move misc functions to utils dir
