+++ app/booth/scrape.py
@@ -156,12 +156,3 @@ def scrape():
 
 
     return booth1_return, booth2_return, booth3_return, booth1_return2, booth2_return2, booth3_return2
-
-def check_banner():
-    banner = open('message.txt', 'r')
-    banner = banner.read()
-
-    if banner == '':
-        banner == False
-        
-    return banner
